requirements.txt
numpy==1.24.3
torch==2.2.0
torchaudio==2.2.0
torchvision==0.17.0
openai-whisper
flask
yt-dlp
werkzeug